Fully move the guts of array_view to span. (#1956)
Now array_view is a synonym for span, rather than the other way around.
Both will work for legacy client apps for some time, but for OIIO internals
we should start favoring span when convenient.

A few minor fixups to make some existing methods and in-class types
match the C++ standard. One change, as dictated by C++20 std::span<>
specification, is that span.size() returns a ptrdiff_t (signed) rather
than a size_t.  This may create some minor build issues with signed vs
unsigned warnings. Something to watch out for.

+++ src/libOpenImageIO/exif.cpp
@@ -254,7 +254,7 @@ tiff_dir_data (const TIFFDirEntry &td, array_view<const uint8_t> data)
     }
     // Long data
     size_t begin = td.tdir_offset;
-    if (begin+len > data.size()) {
+    if (begin+len > size_t(data.size())) {
         // Invalid span -- it is not entirely contained in the data window.
         // Signal error by returning an empty array_view.
         return array_view<const uint8_t>();
